+++ Microsoft.Azure.Cosmos.Encryption/src/Custom/EncryptionProcessor.cs
@@ -572,7 +572,6 @@ namespace Microsoft.Azure.Cosmos.Encryption.Custom
             CancellationToken cancellationToken)
         {
             JObject contentJObj = EncryptionProcessor.BaseSerializer.FromStream<JObject>(content);
-            JArray result = new JArray();
 
             if (!(contentJObj.SelectToken(Constants.DocumentsResourcePropertyName) is JArray documents))
             {
@@ -583,37 +582,23 @@ namespace Microsoft.Azure.Cosmos.Encryption.Custom
             {
                 if (!(value is JObject document))
                 {
-                    result.Add(value);
                     continue;
                 }
 
                 CosmosDiagnosticsContext diagnosticsContext = CosmosDiagnosticsContext.Create(null);
                 using (diagnosticsContext.CreateScope("EncryptionProcessor.DeserializeAndDecryptResponseAsync"))
                 {
-                    (JObject decryptedDocument, DecryptionContext _) = await EncryptionProcessor.DecryptAsync(
+                    await EncryptionProcessor.DecryptAsync(
                         document,
                         encryptor,
                         diagnosticsContext,
                         cancellationToken);
-
-                    result.Add(decryptedDocument);
                 }
             }
 
-            JObject decryptedResponse = new JObject();
-            foreach (JProperty property in contentJObj.Properties())
-            {
-                if (property.Name.Equals(Constants.DocumentsResourcePropertyName))
-                {
-                    decryptedResponse.Add(property.Name, (JToken)result);
-                }
-                else
-                {
-                    decryptedResponse.Add(property.Name, property.Value);
-                }
-            }
-
-            return EncryptionProcessor.BaseSerializer.ToStream(decryptedResponse);
+            // the contents of contentJObj get decrypted in place for MDE algorithm model, and for legacy model _ei property is removed
+            // and corresponding decrypted properties are added back in the documents.
+            return EncryptionProcessor.BaseSerializer.ToStream(contentJObj);
         }
     }
 }