+++ Microsoft.Azure.Cosmos.Encryption/src/EncryptionContainer.cs
@@ -12,7 +12,6 @@ namespace Microsoft.Azure.Cosmos.Encryption
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Cosmos;
-    using Microsoft.Data.Encryption.Cryptography;
     using Newtonsoft.Json.Linq;
 
     internal sealed class EncryptionContainer : Container
@@ -652,7 +651,7 @@ namespace Microsoft.Azure.Cosmos.Encryption
             CosmosDiagnosticsContext diagnosticsContext = CosmosDiagnosticsContext.Create(requestOptions);
             using (diagnosticsContext.CreateScope("PatchItem"))
             {
-                List<PatchOperation> encryptedPatchOperations = await this.PatchItemHelperAsync(
+                List<PatchOperation> encryptedPatchOperations = await this.EncryptPatchOperationsAsync(
                     patchOperations,
                     encryptionSettings,
                     cancellationToken);
@@ -674,7 +673,7 @@ namespace Microsoft.Azure.Cosmos.Encryption
             }
         }
 
-        private async Task<List<PatchOperation>> PatchItemHelperAsync(
+        internal async Task<List<PatchOperation>> EncryptPatchOperationsAsync(
             IReadOnlyList<PatchOperation> patchOperations,
             EncryptionSettings encryptionSettings,
             CancellationToken cancellationToken = default)
@@ -857,15 +856,23 @@ namespace Microsoft.Azure.Cosmos.Encryption
             ReadManyRequestOptions readManyRequestOptions = null,
             CancellationToken cancellationToken = default)
         {
-            throw new NotImplementedException();
+            return this.ReadManyItemsHelperAsync(
+                items,
+                readManyRequestOptions,
+                cancellationToken);
         }
 
-        public override Task<FeedResponse<T>> ReadManyItemsAsync<T>(
+        public override async Task<FeedResponse<T>> ReadManyItemsAsync<T>(
             IReadOnlyList<(string id, PartitionKey partitionKey)> items,
             ReadManyRequestOptions readManyRequestOptions = null,
             CancellationToken cancellationToken = default)
         {
-            throw new NotImplementedException();
+            ResponseMessage responseMessage = await this.ReadManyItemsHelperAsync(
+                items,
+                readManyRequestOptions,
+                cancellationToken);
+
+            return this.ResponseFactory.CreateItemFeedResponse<T>(responseMessage);
         }
 
         public async Task<EncryptionSettings> GetOrUpdateEncryptionSettingsFromCacheAsync(
@@ -877,6 +884,40 @@ namespace Microsoft.Azure.Cosmos.Encryption
                 obsoleteValue: obsoleteEncryptionSettings,
                 singleValueInitFunc: () => EncryptionSettings.CreateAsync(this, cancellationToken),
                 cancellationToken: cancellationToken);
+        }
+
+        internal async Task<Stream> DeserializeAndDecryptResponseAsync(
+           Stream content,
+           EncryptionSettings encryptionSettings,
+           CancellationToken cancellationToken)
+        {
+            if (!encryptionSettings.PropertiesToEncrypt.Any())
+            {
+                return content;
+            }
+
+            JObject contentJObj = EncryptionProcessor.BaseSerializer.FromStream<JObject>(content);
+
+            if (!(contentJObj.SelectToken(Constants.DocumentsResourcePropertyName) is JArray documents))
+            {
+                throw new InvalidOperationException("Feed Response body contract was violated. Feed response did not have an array of Documents. ");
+            }
+
+            foreach (JToken value in documents)
+            {
+                if (value is not JObject document)
+                {
+                    continue;
+                }
+
+                await EncryptionProcessor.DecryptAsync(
+                    document,
+                    encryptionSettings,
+                    cancellationToken);
+            }
+
+            // the contents get decrypted in place by DecryptAsync.
+            return EncryptionProcessor.BaseSerializer.ToStream(contentJObj);
         }
 
         /// <summary>
@@ -926,7 +967,7 @@ namespace Microsoft.Azure.Cosmos.Encryption
 
             ItemRequestOptions clonedRequestOptions = requestOptions;
 
-            // only clone it on the first try.
+            // Clone(once) the request options since we modify it to set AddRequestHeaders to add additional headers.
             if (!isRetry)
             {
                 clonedRequestOptions = GetClonedItemRequestOptions(requestOptions);
@@ -1001,7 +1042,7 @@ namespace Microsoft.Azure.Cosmos.Encryption
 
             ItemRequestOptions clonedRequestOptions = requestOptions;
 
-            // only clone it on the first try.
+            // Clone(once) the request options since we modify it to set AddRequestHeaders to add additional headers.
             if (!isRetry)
             {
                 clonedRequestOptions = GetClonedItemRequestOptions(requestOptions);
@@ -1075,7 +1116,7 @@ namespace Microsoft.Azure.Cosmos.Encryption
 
             ItemRequestOptions clonedRequestOptions = requestOptions;
 
-            // only clone it on the first try.
+            // Clone(once) the request options since we modify it to set AddRequestHeaders to add additional headers.
             if (!isRetry)
             {
                 clonedRequestOptions = GetClonedItemRequestOptions(requestOptions);
@@ -1151,7 +1192,7 @@ namespace Microsoft.Azure.Cosmos.Encryption
 
             ItemRequestOptions clonedRequestOptions = requestOptions;
 
-            // only clone it on the first try.
+            // Clone(once) the request options since we modify it to set AddRequestHeaders to add additional headers.
             if (!isRetry)
             {
                 clonedRequestOptions = GetClonedItemRequestOptions(requestOptions);
@@ -1237,107 +1278,84 @@ namespace Microsoft.Azure.Cosmos.Encryption
 
             foreach (JObject document in documents)
             {
-                try
-                {
-                    JObject decryptedDocument = await EncryptionProcessor.DecryptAsync(
-                        document,
-                        encryptionSettings,
-                        cancellationToken);
+                JObject decryptedDocument = await EncryptionProcessor.DecryptAsync(
+                    document,
+                    encryptionSettings,
+                    cancellationToken);
 
-                    decryptedItems.Add(decryptedDocument.ToObject<T>());
-                }
-
-                // we cannot rely currently on a specific exception, this is due to the fact that the run time issue can be variable,
-                // we can hit issue with either Json serialization say an item was not encrypted but the policy shows it as encrypted,
-                // or we could hit a MicrosoftDataEncryptionException from MDE lib etc.
-                catch (Exception)
-                {
-                    // most likely the encryption policy has changed.
-                    encryptionSettings = await this.GetOrUpdateEncryptionSettingsFromCacheAsync(
-                        obsoleteEncryptionSettings: encryptionSettings,
-                        cancellationToken: cancellationToken);
-
-                    JObject decryptedDocument = await EncryptionProcessor.DecryptAsync(
-                           document,
-                           encryptionSettings,
-                           cancellationToken);
-
-                    decryptedItems.Add(decryptedDocument.ToObject<T>());
-                }
+                decryptedItems.Add(decryptedDocument.ToObject<T>());
             }
 
             return decryptedItems;
         }
 
-        internal async Task<Stream> DeserializeAndDecryptResponseAsync(
-            Stream content,
-            EncryptionSettings encryptionSettings,
-            CancellationToken cancellationToken)
+        private async Task<ResponseMessage> ReadManyItemsHelperAsync(
+            IReadOnlyList<(string id, PartitionKey partitionKey)> items,
+            ReadManyRequestOptions readManyRequestOptions = null,
+            CancellationToken cancellationToken = default,
+            bool isRetry = false)
         {
+            EncryptionSettings encryptionSettings = await this.GetOrUpdateEncryptionSettingsFromCacheAsync(
+               obsoleteEncryptionSettings: null,
+               cancellationToken: cancellationToken);
+
             if (!encryptionSettings.PropertiesToEncrypt.Any())
             {
-                return content;
+                return await this.container.ReadManyItemsStreamAsync(
+                    items,
+                    readManyRequestOptions,
+                    cancellationToken);
             }
 
-            JObject contentJObj = EncryptionProcessor.BaseSerializer.FromStream<JObject>(content);
-            JArray results = new JArray();
+            ReadManyRequestOptions clonedRequestOptions = readManyRequestOptions;
 
-            if (!(contentJObj.SelectToken(Constants.DocumentsResourcePropertyName) is JArray documents))
+            // Clone(once) the request options since we modify it to set AddRequestHeaders to add additional headers.
+            if (!isRetry)
             {
-                throw new InvalidOperationException("Feed Response body contract was violated. Feed response did not have an array of Documents. ");
-            }
-
-            foreach (JToken value in documents)
-            {
-                if (value is not JObject document)
+                if (readManyRequestOptions != null)
                 {
-                    results.Add(value);
-                    continue;
-                }
-
-                try
-                {
-                    JObject decryptedDocument = await EncryptionProcessor.DecryptAsync(
-                        document,
-                        encryptionSettings,
-                        cancellationToken);
-
-                    results.Add(decryptedDocument);
-                }
-
-                // we cannot rely currently on a specific exception, this is due to the fact that the run time issue can be variable,
-                // we can hit issue with either Json serialization say an item was not encrypted but the policy shows it as encrypted,
-                // or we could hit a MicrosoftDataEncryptionException from MDE lib etc.
-                catch (Exception)
-                {
-                    // most likely the encryption policy has changed.
-                    encryptionSettings = await this.GetOrUpdateEncryptionSettingsFromCacheAsync(
-                        obsoleteEncryptionSettings: encryptionSettings,
-                        cancellationToken: cancellationToken);
-
-                    JObject decryptedDocument = await EncryptionProcessor.DecryptAsync(
-                        document,
-                        encryptionSettings,
-                        cancellationToken);
-
-                    results.Add(decryptedDocument);
-                }
-            }
-
-            JObject decryptedResponse = new JObject();
-            foreach (JProperty property in contentJObj.Properties())
-            {
-                if (property.Name.Equals(Constants.DocumentsResourcePropertyName))
-                {
-                    decryptedResponse.Add(property.Name, (JToken)results);
+                    clonedRequestOptions = (ReadManyRequestOptions)readManyRequestOptions.ShallowCopy();
                 }
                 else
                 {
-                    decryptedResponse.Add(property.Name, property.Value);
+                    clonedRequestOptions = new ReadManyRequestOptions();
                 }
             }
 
-            return EncryptionProcessor.BaseSerializer.ToStream(decryptedResponse);
+            encryptionSettings.SetRequestHeaders(clonedRequestOptions);
+
+            ResponseMessage responseMessage = await this.container.ReadManyItemsStreamAsync(
+                items,
+                clonedRequestOptions,
+                cancellationToken);
+
+            if (!isRetry &&
+                responseMessage.StatusCode == HttpStatusCode.BadRequest &&
+                string.Equals(responseMessage.Headers.Get(Constants.SubStatusHeader), Constants.IncorrectContainerRidSubStatus))
+            {
+                // get the latest encryption settings.
+                await this.GetOrUpdateEncryptionSettingsFromCacheAsync(
+                    obsoleteEncryptionSettings: encryptionSettings,
+                    cancellationToken: cancellationToken);
+
+                return await this.ReadManyItemsHelperAsync(
+                    items,
+                    clonedRequestOptions,
+                    cancellationToken,
+                    isRetry: true);
+            }
+
+            if (responseMessage.IsSuccessStatusCode && responseMessage.Content != null)
+            {
+                Stream decryptedContent = await this.DeserializeAndDecryptResponseAsync(
+                    responseMessage.Content,
+                    encryptionSettings,
+                    cancellationToken);
+
+                return new DecryptedResponseMessage(responseMessage, decryptedContent);
+            }
+
+            return responseMessage;
         }
     }
 }